requirements.txt
django~=3.1.7
djangorestframework~=3.12.2
django-cors-headers
channels
django-colorfield